+++ install.sh
@@ -2,7 +2,7 @@
 
 ## ASSIGN VARIABLE
 
-project_version="2.8.8"
+project_version="2.8.9"
 
 printf -v date_time '%(%Y-%m-%d_%H-%M-%S)T' -1
 log_file="adsb-installer_${date_time}.log"